refactor(router): 이름 통일화 및 url 변경 / DP-96
- 피그마의 사이드바 툴팁 기준으로 변경
