+++ src/router/routeTree.ts
@@ -6,16 +6,12 @@ import { signUpRoute } from './routes/auth/sign-up';
 import { findIdRoute } from './routes/auth/find-id';
 import { findPasswordRoute } from './routes/auth/find-password';
 import { mainLayoutRoute } from './routes/main/main-layout';
-import { myRepositoriesRoute } from './routes/main/my-repositories';
-import { SharedRepositoriesRoute } from './routes/main/shared-repositories';
-import { SharedMeRepositoriesRoute } from './routes/main/shared-me-repositories';
+import { PrivateRepoPageRoute } from './routes/main/private-repo';
+import { SharedByMeRepoRoute } from './routes/main/shared-by-me-repo';
+import { SharedWithMeRepoRoute } from './routes/main/shared-with-me-repo';
 
 export const routeTree = rootRoute.addChildren([
   landingRoute,
   authLayoutRoute.addChildren([signInRoute, signUpRoute, findIdRoute, findPasswordRoute]),
-  mainLayoutRoute.addChildren([
-    myRepositoriesRoute,
-    SharedRepositoriesRoute,
-    SharedMeRepositoriesRoute,
-  ]),
+  mainLayoutRoute.addChildren([PrivateRepoPageRoute, SharedByMeRepoRoute, SharedWithMeRepoRoute]),
 ]);
