--- src/router/routes/main/my-repositories.tsx
@@ -1,10 +1,0 @@
-import { createRoute } from '@tanstack/react-router';
-import { mainLayoutRoute } from '@/router/routes/main/main-layout';
-
-import MyRepositoriesPage from '@/pages/Main/MyRepositoriesPage/MyRepositoriesPage';
-
-export const myRepositoriesRoute = createRoute({
-  getParentRoute: () => mainLayoutRoute,
-  path: '/main/my-repo',
-  component: MyRepositoriesPage,
-});
