--- src/router/routes/main/shared-repositories.tsx
@@ -1,10 +1,0 @@
-import { createRoute } from '@tanstack/react-router';
-import { mainLayoutRoute } from '@/router/routes/main/main-layout';
-
-import SharedRepositoriesPage from '@/pages/Main/SharedRepositoryPage/SharedRepositoriesPage';
-
-export const SharedRepositoriesRoute = createRoute({
-  getParentRoute: () => mainLayoutRoute,
-  path: '/main/shared-repo',
-  component: SharedRepositoriesPage,
-});
