+++ src/router/routes/main/shared-with-me-repo.tsx
@@ -1,0 +1,10 @@
+import { createRoute } from '@tanstack/react-router';
+import { mainLayoutRoute } from '@/router/routes/main/main-layout';
+
+import SharedWithMeRepoPage from '@/pages/Main/SharedWithMeRepoPage/SharedWithMeRepoPage';
+
+export const SharedWithMeRepoRoute = createRoute({
+  getParentRoute: () => mainLayoutRoute,
+  path: '/main/shared-with-me-repo',
+  component: SharedWithMeRepoPage,
+});
